Updated Scientific Notation & Mass.
Updated the project to properly display Scientific Notation in the request and updated the way information would be stored in backend.
