+++ src/main/java/com/solsorg/solarsim/model/Star.java
@@ -2,33 +2,20 @@ package com.solsorg.solarsim.model;
 
 import com.solsorg.solarsim.util.math.ScientificNotation;
 import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 
 @Entity
 public class Star extends SpaceEntity{
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
     private String name;
-
     private double radius;
 
-    public Star(double x, double y, double z, ScientificNotation mass){
+    public Star(double x, double y, double z, ScientificNotation mass, String name){
         super(x, y, z, mass);
+
+        this.name = name;
     }
 
     public Star() {
         super();
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
     }
 
     public String getName() {
